requirements.txt
asn1crypto==0.24.0
boto3==1.9.200
botocore==1.12.200
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
docutils==0.14
ecdsa==0.13.2
future==0.17.1
idna==2.8
jmespath==0.9.4
psycopg2==2.8.3
pyasn1==0.4.6
pycparser==2.19
python-dateutil==2.8.0
python-jose==3.0.1
python-telegram-bot==11.1.0
pytz==2019.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
